requirements.txt
txtai
datasets
openai
python-dotenv
tiktoken